# contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: MaxMarket API
  version: "1.0.2"
  description: |
    MaxMarket B2B wholesale platform API.
    Base URL: /api/v1

    **Correlation ID:** Every request MAY include an `X-Correlation-ID` header (UUID).
    If provided, the server echoes it in the response. If absent, the server generates one.
    Every response carries `X-Correlation-ID` in the response header.
    Every error response includes `correlationId` in the body.

    **Error envelope:** All error responses use the `ErrorResponse` schema.

    **Public catalog:** `GET /catalog/products` and `GET /catalog/products/{id}` require no
    authentication and return `PublicProduct` — price fields are entirely omitted.
    Authenticated requests to those same endpoints return `Product` with role-appropriate
    price fields resolved server-side.

    **Optimistic locking:** `POST /orders/{id}/approve` AND `PUT /orders/{id}` (when editing
    an Approved order) both require the current `versionLock` value to prevent concurrent
    conflicting writes.

    **Draft order pricing:** At Draft creation, `groupDiscount = 0` and `finalPrice = basePrice`
    (captured from current `pricePerUnit`). Both values are recalculated on submission from
    the then-current client group discount rule (DN-PRICE-02).

    **Decision Needed (open):**
    - DA-05: Can an Admin change a variant's SKU if non-Draft orders reference it?
...
      Current spec assumes URL-only (CDN pre-upload pattern).

...

...
servers:
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── HEADERS ────────────────────────────────────────────────────────────────
  headers:
    X-Correlation-ID:
      description: |
        UUID for request tracing. Echoed from request X-Correlation-ID header if provided,
        otherwise server-generated. Present on every response.
      schema:
        type: string
        format: uuid

  # ─── PARAMETERS ─────────────────────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    CorrelationIdRequestHeader:
      name: X-Correlation-ID
      in: header
      required: false
      description: Optional client-provided correlation UUID. Server echoes it in the response header.
      schema:
        type: string
        format: uuid

  # ─── SCHEMAS ────────────────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorCode, message, correlationId]
      properties:
        errorCode:
          type: string
          description: |
            Machine-readable error code. Known values include:
            INSUFFICIENT_STOCK, ORDER_NOT_EDITABLE, OPTIMISTIC_LOCK_CONFLICT,
            STOCK_BELOW_RESERVED, VALIDATION_ERROR, TOKEN_EXPIRED, TOKEN_MISSING,
            FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED
          example: INSUFFICIENT_STOCK
        message:
          type: string
          description: Human-readable error message.
        details:
          nullable: true
          description: |
            Structured error detail. Schema depends on errorCode:
            - INSUFFICIENT_STOCK: array of InsufficientStockDetail
            - VALIDATION_ERROR: array of Zod validation errors
            - All others: null or plain string
        correlationId:
          type: string
          format: uuid
...
    PaginationMeta:
      type: object
      required: [page, pageSize, totalCount, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer

    # --- Common ---
...
    ClientGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        discountType:
          type: string
          enum: [fixed, percentage]
        discountValue:
          type: number
          format: double

    # --- Catalog (Public — NO price fields) ---
...
    WarehouseStock:
      type: object
      properties:
        variantId:
          type: string
          format: uuid
        sku:
          type: string
        productName:
          type: string
          nullable: true
          description: Display name of the product
        warehouseId:
          type: string
          format: uuid
        availableQty:
          type: integer
        reservedQty:
          type: integer

    # --- Orders ---
...
security:
  - bearerAuth: []
paths:
...
  /client-groups:
    get:
      tags: [Client Groups]
      summary: List client groups
      description: "Roles: super_admin, admin, manager, agent. Note: client role cannot access."
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      responses:
        '200':
          description: Paginated client groups
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClientGroup'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
    post:
      tags: [Client Groups]
      summary: Create client group
      description: "Roles: super_admin, admin"
      parameters:
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, discountType, discountValue]
              properties:
                name:
                  type: string
                discountType:
                  type: string
                  enum: [fixed, percentage]
                discountValue:
                  type: number
                  minimum: 0
      responses:
        '201':
          description: Created
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientGroup'
        '422':
          description: Validation error
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /client-groups/{id}:
    put:
      tags: [Client Groups]
      summary: Update client group
      description: |
        Roles: super_admin, admin.
        Changes affect new orders and existing Draft orders on next submission (DN-PRICE-02).
        Does NOT affect Approved or Fulfilled orders.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                discountType:
                  type: string
                  enum: [fixed, percentage]
                discountValue:
                  type: number
                  minimum: 0
      responses:
        '200':
          description: Updated
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientGroup'
        '404':
          description: Not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Client Groups]
      summary: Delete client group
      description: "Roles: super_admin, admin. Blocked if any clients are assigned to the group."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      responses:
        '200':
          description: Deleted
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
        '404':
          description: Not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Clients are assigned to this group
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── CATALOG ─────────────────────────────────────────────────────────────────
...
  /inventory/stock:
    get:
      tags: [Inventory]
      summary: List stock levels
      description: "Roles: super_admin, admin, manager, agent. Client role cannot access stock data."
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
        - name: warehouseId
          in: query
          schema:
            type: string
            format: uuid
        - name: variantId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated stock records
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WarehouseStock'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '403':
          description: Forbidden
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory/stock/adjust:
    put:
      tags: [Inventory]
      summary: Adjust available stock quantity
      description: |
        Roles: super_admin, admin only.
        Sets `available_qty` to the specified value. Cannot be set below `reserved_qty`
        (PRD §5.3, CTO_PACKET §5). All adjustments are written to audit log (stock.adjusted event).
        Error code STOCK_BELOW_RESERVED returned if constraint violated.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [warehouseId, variantId, newAvailableQty, reason]
              properties:
                warehouseId:
                  type: string
                  format: uuid
                variantId:
                  type: string
                  format: uuid
                newAvailableQty:
                  type: integer
                  minimum: 0
                reason:
                  type: string
                  description: Required reason for audit trail.
      responses:
        '200':
          description: Stock updated
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseStock'
        '422':
          description: |
            Constraint violation. errorCode = STOCK_BELOW_RESERVED if newAvailableQty < reservedQty.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: STOCK_BELOW_RESERVED
                message: "Cannot set available_qty to 5: reserved_qty is 12. Release reservations first."
                details:
                  requestedAvailableQty: 5
                  currentReservedQty: 12
                correlationId: "550e8400-e29b-41d4-a716-446655440000"
        '404':
          description: Variant or warehouse not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── ORDERS ──────────────────────────────────────────────────────────────────
...